Move pkg/storage/filesystem/{watch,fileevents}. Make the Emitter interface a bit simpler, yay!
